Trait vortex_buffer::io_buf::IoBuf

pub unsafe trait IoBuf: Unpin + 'static {
    // Required methods
    fn read_ptr(&self) -> *const u8;
    fn bytes_init(&self) -> usize;
    fn as_slice(&self) -> &[u8] ;

    // Provided method
    fn slice_owned(self, range: Range<usize>) -> Slice<Self>
       where Self: Sized { ... }
}

Trait for types that can provide a readonly byte buffer interface to I/O frameworks.

Safety

The type must support contiguous raw memory access via pointer, such as Vec or [u8].

Required Methods

fn read_ptr(&self) -> *const u8

Returns a raw pointer to the vector’s buffer.

fn bytes_init(&self) -> usize

Number of initialized bytes.

fn as_slice(&self) -> &[u8]

Access the buffer as a byte slice

Provided Methods

fn slice_owned(self, range: Range<usize>) -> Slice<Self>
where Self: Sized,

Access the buffer as a byte slice with begin and end indices